refactor addIntroOutro into separate files
