+++ functions/src/addIntroOutro/CancelToken.ts
@@ -1,0 +1,11 @@
+export class CancelToken {
+    private shouldCancel = false;
+  
+    cancel = () => {
+      this.shouldCancel = true;
+    }
+  
+    get isCancellationRequested() {
+      return this.shouldCancel;
+    }
+  }
